Add files via upload
Twitter is one of the most popular social media sites for individuals to engage, share content, and comment on a variety of topics, including health difficulties.The world is currently suffering from the COVID-19 pandemic, and people are starting to tweet about it in droves. This vast amount of data could be useful, but it must be analyzed in the context of the COVID-19 epidemic, both positively and negatively.

As a result, sentiment analysis of Twitter streaming data can help spread medical information about the coronavirus outbreak.
 
Twitter APIs have been investigated for consuming streaming data in order to make the proposed system's data collection stage a reality. The Search API and the Stream API are two of Twitter's streaming APIs.


Advantages of Sentiment analysis on Business application of analysis
Brand revitalization
Elimination of guess work by using insight rich information
Timely decision making using automated insights
Enhancing customer or user experience